Fix: "Add Collectionview in SpecificPost (Changed UI)"
